add bigint

+++ app/database.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy import Column, Integer, String, DateTime, Text, func, Index, Boolean
+from sqlalchemy import BigInteger,Column, Integer, String, DateTime, Text, func, Index, Boolean
 from config import engine
 
 # Базовый класс
@@ -10,7 +10,7 @@ class Base(DeclarativeBase):
 class User(Base):
     __tablename__ = "users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     full_name = Column(String, nullable=True)
     username = Column(String, nullable=True)
     free_messages = Column(Integer, default=10)  # Количество бесплатных сообщений
@@ -24,7 +24,7 @@ class ChatHistory(Base):
     __tablename__ = "chat_history"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     user_message = Column(Text, nullable=False)
     bot_response = Column(Text, nullable=False)
     timestamp = Column(DateTime, server_default=func.now())
